+++ android/src/main/java/com/backgroundlocation/database/TrackingStateEntity.kt
@@ -25,6 +25,13 @@ data class TrackingStateEntity(
   val notificationText: String? = null,
   val notificationChannelName: String? = null,
   val notificationPriority: String? = null,
-  val foregroundOnly: Boolean? = null
+  val foregroundOnly: Boolean? = null,
+  val notificationSmallIcon: String? = null,
+  val notificationColor: String? = null,
+  val notificationShowTimestamp: Boolean? = null,
+  val notificationActions: String? = null,
+  val notificationLargeIcon: String? = null,
+  val notificationSubtext: String? = null,
+  val notificationChannelId: String? = null
 )